generalize copy clipboard functionality
