+++ src/components/CopyButton.js
@@ -5,6 +5,7 @@ import { css } from '@emotion/core';
 import Icon from '@leafygreen-ui/icon';
 import IconButton from '@leafygreen-ui/icon-button';
 import { uiColors } from '@leafygreen-ui/palette';
+import useCopyClipboard from '../hooks/useCopyClipboard';
 
 const getCopyButtonStyle = (copied) => {
   const baseStyle = css`
@@ -36,25 +37,7 @@ const CopyButton = ({ contents, onCopy }) => {
   const [copied, setCopied] = useState(false);
   const [buttonNode, setButtonNode] = useState(null);
 
-  useEffect(() => {
-    if (!buttonNode) {
-      return;
-    }
-
-    const clipboard = new ClipboardJS(buttonNode, {
-      text: () => contents,
-    });
-
-    if (copied) {
-      const timeoutId = setTimeout(() => {
-        setCopied(false);
-      }, 1500);
-
-      return () => clearTimeout(timeoutId);
-    }
-
-    return () => clipboard.destroy();
-  }, [buttonNode, contents, copied]);
+  useCopyClipboard(copied, setCopied, buttonNode, contents);
 
   const handleClick = (e) => {
     e.preventDefault();
