+++ src/hooks/useCopyClipboard.js
@@ -1,0 +1,26 @@
+import ClipboardJS from 'clipboard';
+import { useEffect } from 'react';
+
+const useCopyClipboard = (copied, setCopied, component, contents) => {
+  useEffect(() => {
+    if (!component) {
+      return;
+    }
+
+    const clipboard = new ClipboardJS(component, {
+      text: () => contents,
+    });
+
+    if (copied) {
+      const timeoutId = setTimeout(() => {
+        setCopied(false);
+      }, 1500);
+
+      return () => clearTimeout(timeoutId);
+    }
+
+    return () => clipboard.destroy();
+  }, [component, contents, copied, setCopied]);
+};
+
+export default useCopyClipboard;
